requirements.txt
# requirements.txt

# FastAPI and dependencies
fastapi==0.105.0
uvicorn==0.23.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# AI/ML
openai==1.3.5

# Testing
pytest==7.4.3
httpx==0.25.2